refactor(utils): move ui_helpers.py to app_frame handlers
Only consumers are widgets/frames/app_frame/handlers/app_events.py and
app_frame.py. Colocate with its callers.

Refs #441
