+++ widgets/frames/app_frame/handlers/app_events.py
@@ -13,7 +13,7 @@ from utils.card_data import CardDataManager
 from utils.constants import LOGS_DIR
 from utils.deck import sanitize_filename
 from utils.deck_results_filter import _classify_event_type, _normalize_date, filter_decks
-from utils.ui_helpers import open_child_window, widget_exists
+from widgets.frames.app_frame.handlers.ui_helpers import open_child_window, widget_exists
 from widgets.dialogs.feedback_dialog import show_feedback_dialog
 from widgets.frames.identify_opponent import MTGOpponentDeckSpy
 from widgets.frames.match_history import MatchHistoryFrame
